requirements.txt
tensorflow-cpu==2.15.0
numpy==1.24.3
pillow==9.5.0
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
h5py==3.1.0
gunicorn==20.1.0